E2E tests split
