+++ ts/__e2e__/login.e2e.ts
@@ -1,9 +1,7 @@
 import { loginWithSPID } from "./utils";
 
-describe("User Login using SPID", () => {
-  describe("when the user never logged in before", () => {
-    it("should let the user log in with SPID", async () => {
-      await loginWithSPID();
-    });
+describe("User Login using SPID, when the user never logged in before", () => {
+  it("should let the user log in with SPID", async () => {
+    await loginWithSPID();
   });
 });
